refactor styling, function
