--- src/app/onboarding/basic-information/styles.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const LineDiv = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 2rem; // 40 px
-  width: 100%;
-`;
